Core functionality

+++ main.py
@@ -3,11 +3,11 @@
 from flask import Flask, render_template, request, flash, redirect, url_for, session
 from flask_mysqldb import MySQL
 import os
-#from passlib.hash import pbkdf2_sha256
 from passlib.hash import sha256_crypt
 from FrostCryption import *
-from PassMaker import *
+from PassMaker import get_pass, Secret_Key
 import pyperclip
+#from SendEmails import *
 
 
 # CONFIG
@@ -21,7 +21,7 @@ app.config['MYSQL_DB'] = 'PassManager'
 app.config['MYSQL_CURSORCLASS'] = 'DictCursor'
 app.config['SEND_FILE_MAX_AGE_DEFAULT'] = 0
 app.secret_key = os.urandom(24)
-etext = '\xb1<\x85\x9f\x18\xe0b\xf3AJb\xce\x94\xd4f\xea\x19@\xc3G\xc6\x8a\xf0'
+
 
 # intialising MySQL
 mysql = MySQL()
@@ -50,11 +50,12 @@ def CPasscodes():
         passcode = get_pass(passlen)
         pyperclip.copy(passcode)
         flash('Password Copied', 'success')
+        etext = get_pass(128)
         passcode = FrostCrypt(str(passcode), etext)
-
+        Etext = FrostCrypt(str(etext), str(session['user']))
         cur = mysql.connection.cursor()
-        cur.execute("INSERT INTO main (url, username, passcode, uid)VALUES ('{}','{}','{}',{})".format(
-            url, username, passcode, session['user']))
+        cur.execute("INSERT INTO main (url, username, passcode, uid, etext)VALUES ('{}','{}','{}',{},'{}')".format(
+            url, username, passcode, session['user'], Etext))
         mysql.connection.commit()
         cur.close()
 
@@ -65,6 +66,38 @@ def CPasscodes():
     return render_template("CPasscodes.html", data=data)
 
 
+@app.route('/CPasscodes1', methods=['GET', 'POST'])
+def CPasscodes1():
+    if 'user' not in session:
+        flash("Please Login To Add Password", 'danger')
+        return redirect('/login')
+    if request.method == 'POST':
+        url = request.form['URL']
+        username = request.form['Username']
+        passcode = request.form['passcode']
+
+        if username == "":
+            username = 'N/A'
+
+        pyperclip.copy(passcode)
+        flash('Password Copied', 'success')
+        etext = get_pass(128)
+        passcode = FrostCrypt(str(passcode), etext)
+        Etext = FrostCrypt(str(etext), str(session['user']))
+        cur = mysql.connection.cursor()
+        cur.execute("INSERT INTO main (url, username, passcode, uid, etext)VALUES ('{}','{}','{}',{},'{}')".format(
+            url, username, passcode, session['user'], Etext))
+        mysql.connection.commit()
+        cur.close()
+
+    cur = mysql.connection.cursor()
+    cur.execute(
+        "select name, username from user where uid = {}".format(session['user']))
+    data = cur.fetchone()
+    cur.close()
+    return render_template("CPasscodes1.html", data=data)
+
+
 @app.route('/mypasscodes', methods=['GET', 'POST'])
 def mypasscodes():
     if 'user' not in session:
@@ -73,14 +106,15 @@ def mypasscodes():
     if request.method == 'POST':
         url = str(request.form['URL'])
         cur = mysql.connection.cursor()
-        cur.execute("select `url`, `username`, `primarykey` from main where `url` REGEXP '{}' and uid = {} ORDER BY date_time DESC".format(
+        cur.execute("select `url`, `username`, `primarykey`,`date_time` from main where `url` REGEXP '{}' and uid = {} order by `date_time` DESC".format(
             url, session['user']))
         data = cur.fetchall()
         cur.close()
         return render_template('mypasscodes.html', data=data)
 
     cur = mysql.connection.cursor()
-    cur.execute("Select * from main where uid = {}".format(session['user']))
+    cur.execute(
+        "Select * from main where uid = {} order by date_time DESC".format(session['user']))
     data = cur.fetchall()
 
     cur.close()
@@ -94,18 +128,25 @@ def copy(primarykey):
         flash("Please Login", 'danger')
         return redirect('/login')
     cur = mysql.connection.cursor()
-    cur.execute(
-        "select `passcode` from main where `primarykey` = '{}' ORDER BY date_time DESC".format(primarykey))
-    data = cur.fetchone()
-    cur.close()
-    passcode = FrostDCrypt(data['passcode'], etext)
 
-    try:
-        pyperclip.copy(passcode)
-    except(PermissionError):
-        print("There is an error!")
-    flash('Password Copied', 'success')
-    return redirect('/mypasscodes')
+    cur.execute(
+        "select `passcode`,`etext` from main where `primarykey` = '{}' and `uid` = '{}' ORDER BY date_time DESC".format(primarykey,session['user']))
+    data = cur.fetchone()  
+    cur.close() 
+    if data:
+        etext = FrostDCrypt(data['etext'], str(session['user']))
+        passcode = FrostDCrypt(data['passcode'], etext)
+        try:
+            pyperclip.copy(passcode)
+        except(PermissionError):
+            print("There is an error!")
+        flash('Password Copied', 'success')
+        return redirect('/mypasscodes')
+    else: 
+        return ('Ops Something Went Wrong')
+
+    
+    
 
 
 @app.route('/delete/<primarykey>')
@@ -114,7 +155,7 @@ def delete(primarykey):
         flash("Please Login", 'danger')
         return redirect('/login')
     cur = mysql.connection.cursor()
-    cur.execute("delete from main where primarykey = {}".format(primarykey))
+    cur.execute("delete from main where primarykey = {} and `uid` = {}".format(primarykey,session['user']))
     mysql.connection.commit()
     cur.close()
     flash('Password Deleted', 'danger')
@@ -142,15 +183,16 @@ def register():
         password = sha256_crypt.encrypt(str(request.form['password']))
         name = request.form['name']
         email = request.form['email']
-
+        private_key = Secret_Key()
+        SecKey = sha256_crypt.encrypt(str(private_key))
         cur = mysql.connection.cursor()
-        cur.execute("INSERT INTO user(name, email, password, username) VALUES(%s, %s, %s, %s)",
-                    (name, email, password, username))
+        cur.execute("INSERT INTO user(name, email, password, username, privatekeys) VALUES(%s, %s, %s, %s, %s)",
+                    (name, email, password, username, SecKey))
         mysql.connection.commit()
 
         cur.close()
         flash('You are successfully Registered', 'success')
-        return redirect('/login')
+        return render_template('SecretKeys.html', SecKey = private_key)
 
     return render_template("Register.html")
 
@@ -173,18 +215,63 @@ def login():
         if data:
             password = data['password']
             uid = data["uid"]
-            if sha256_crypt.verify(passcode,password):
+            if sha256_crypt.verify(passcode, password):
                 session['user'] = uid
                 flash('Successfully logged in', 'success')
                 return redirect('CPasscodes')
             else:
-                flash('Invalid Log In','danger')
+                flash('Invalid Log In', 'danger')
         else:
             flash('User not Found', 'danger')
     return render_template('Login.html')
 
-# LOGOUT ROUTE
 
+# Forgot Password
+@app.route('/forgot_password', methods=['GET', 'POST'])
+def forgot_password():
+    if 'user' in session:
+        return redirect('/')
+    else:
+        if request.method == 'POST':
+            SecKey = request.form['key']
+            username = request.form['username']
+            cur = mysql.connection.cursor()
+            cur.execute(
+                "select uid,privatekeys from user where username = %s", [username])
+            data = cur.fetchone()
+            cur.close()
+            if data:
+                privatekeys = data['privatekeys']
+                uid = data["uid"]
+                if sha256_crypt.verify(SecKey, privatekeys):
+                    session['user'] = uid
+                    flash('Successfully logged in', 'success')
+                    return redirect('/ChangePassword')
+                else:
+                    flash('Invalid Private Key', 'danger')
+            else:
+                flash('User Not Found', 'danger') 
+        return render_template('forgot_password.html')
+
+@app.route('/ChangePassword', methods=['GET','POST'])
+def ChangePassword():
+    if "user" not in session:
+        flash('Ops! something went wrong', 'danger')
+        return redirect('/login')
+    if request.method == 'POST':
+        password = sha256_crypt.encrypt(str(request.form['password']))
+        private_key = Secret_Key()
+        SecKey = sha256_crypt.encrypt(str(private_key))
+        cur = mysql.connection.cursor()
+        cur.execute("update user set password = %s,privatekeys = %s where uid = %s",(str(password), SecKey,session['user']))
+        mysql.connection.commit()
+        cur.close()
+        flash('Password changed','success')
+        return render_template('SecretKeys.html', SecKey = private_key)
+    return render_template('ChangePassword.html')
+
+
+# LOGOUT ROUTE
 
 @app.route('/logout')
 def logout():
